+++ apps/api/src/v1/routes/coe.ts
@@ -1,9 +1,7 @@
-import { CACHE_TTL } from "@api/config";
 import db from "@api/config/db";
-import redis from "@api/config/redis";
 import { getUniqueMonths } from "@api/lib/getUniqueMonths";
 import { groupMonthsByYear } from "@api/lib/groupMonthsByYear";
-import { type COE, COEQuerySchema, MonthsQuerySchema } from "@api/schemas";
+import { COEQuerySchema, MonthsQuerySchema } from "@api/schemas";
 import { zValidator } from "@hono/zod-validator";
 import { coe, coePQP } from "@sgcarstrends/schema";
 import { and, asc, desc, eq, gte, inArray, lte, max } from "drizzle-orm";
@@ -14,13 +12,6 @@ const app = new Hono();
 app.get("/", zValidator("query", COEQuerySchema), async (c) => {
   const query = c.req.query();
   const { month, from, to } = query;
-
-  const CACHE_KEY = `coe:${JSON.stringify(query)}`;
-
-  const cachedData = await redis.get<COE[]>(CACHE_KEY);
-  if (cachedData) {
-    return c.json(cachedData);
-  }
 
   const filters = [
     month && eq(coe.month, month),
@@ -33,8 +24,6 @@ app.get("/", zValidator("query", COEQuerySchema), async (c) => {
     .from(coe)
     .where(and(...filters))
     .orderBy(desc(coe.month), asc(coe.bidding_no), asc(coe.vehicle_class));
-
-  await redis.set(CACHE_KEY, results, { ex: CACHE_TTL });
 
   return c.json(results);
 });
@@ -51,13 +40,6 @@ app.get("/months", zValidator("query", MonthsQuerySchema), async (c) => {
 });
 
 app.get("/latest", async (c) => {
-  const CACHE_KEY = "coe:latest";
-
-  const cachedData = await redis.get<COE[]>(CACHE_KEY);
-  if (cachedData) {
-    return c.json(cachedData);
-  }
-
   const [{ latestMonth }] = await db
     .select({ latestMonth: max(coe.month) })
     .from(coe);
@@ -78,19 +60,10 @@ app.get("/latest", async (c) => {
     )
     .orderBy(desc(coe.bidding_no), asc(coe.vehicle_class));
 
-  await redis.set(CACHE_KEY, results, { ex: CACHE_TTL });
-
   return c.json(results);
 });
 
 app.get("/pqp", async (c) => {
-  const CACHE_KEY = "coe:pqp";
-
-  const cachedData = await redis.get(CACHE_KEY);
-  if (cachedData) {
-    return c.json(cachedData);
-  }
-
   const results = await db
     .select()
     .from(coePQP)
@@ -103,8 +76,6 @@ app.get("/pqp", async (c) => {
     grouped[month][vehicle_class] = pqp;
     return grouped;
   }, {});
-
-  await redis.set(CACHE_KEY, pqpRates, { ex: CACHE_TTL });
 
   return c.json(pqpRates);
 });