+++ apps/api/src/v1/routes/makes.ts
@@ -1,8 +1,5 @@
-import { CACHE_TTL } from "@api/config";
 import db from "@api/config/db";
-import redis from "@api/config/redis";
 import { MakeParamSchema, MakeQuerySchema } from "@api/schemas";
-import type { Make } from "@api/types";
 import { zValidator } from "@hono/zod-validator";
 import { cars } from "@sgcarstrends/schema";
 import { and, asc, desc, eq, ilike } from "drizzle-orm";
@@ -11,21 +8,11 @@ import { Hono } from "hono";
 const app = new Hono();
 
 app.get("/", async (c) => {
-  const CACHE_KEY = "makes";
-
-  let makes = await redis.zrange<Make[]>(CACHE_KEY, 0, -1);
-  if (makes.length === 0) {
-    makes = await db
-      .selectDistinct({ make: cars.make })
-      .from(cars)
-      .orderBy(asc(cars.make))
-      .then((res) => res.map(({ make }) => make));
-
-    for (const make of makes) {
-      await redis.zadd<Make>(CACHE_KEY, { score: 0, member: make });
-    }
-    await redis.expire(CACHE_KEY, CACHE_TTL);
-  }
+  const makes = await db
+    .selectDistinct({ make: cars.make })
+    .from(cars)
+    .orderBy(asc(cars.make))
+    .then((res) => res.map(({ make }) => make));
 
   return c.json(makes);
 });
@@ -41,13 +28,6 @@ app.get(
       const query = c.req.valid("query");
       const { month, fuel_type, vehicle_type } = query;
 
-      const CACHE_KEY = `make:${make}:${JSON.stringify(query)}`;
-
-      const cachedData = await redis.get(CACHE_KEY);
-      if (cachedData) {
-        return c.json(cachedData);
-      }
-
       const filters = [
         ilike(cars.make, make.split("-").join("%")),
         month && eq(cars.month, month),
@@ -61,8 +41,6 @@ app.get(
         .from(cars)
         .where(and(...filters))
         .orderBy(desc(cars.month), asc(cars.fuel_type), asc(cars.vehicle_type));
-
-      await redis.set(CACHE_KEY, JSON.stringify(results), { ex: 86400 });
 
       return c.json(results);
     } catch (e) {